Update Task.jsx to improve task status handling
The update includes changes to Task.jsx that improves the way task status is handled. The status of a task is now handled with defined statuses (NOTDONE, WORKING, DONE) rather than arbitrary number values. Task information is now updated using the configStore and updateActiveEvent service. The code also includes updates to related dependencies in the project.