Add utility function for environment variable typecasting
A new function `parseEnvToBoolean` has been introduced in `varCasting.js` to handle the typecasting of environment variables. This function is now used across the codebase to convert `VITE_MOCK` to a Boolean value. Also, adjustments have been made to allow drag and drop functionality in `AssestmentCard.jsx` and the `.env` file is now ignored by Git.